+++ src/main/java/pl/peterwolf/echoesinink/structure/AbandonedPrintshopPiece.java
@@ -6,6 +6,8 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.StructureManager;
 import net.minecraft.world.level.WorldGenLevel;
@@ -21,8 +23,12 @@ import net.minecraft.world.level.levelgen.structure.pieces.StructurePieceSeriali
 import net.minecraft.world.level.storage.loot.LootTable;
 import pl.peterwolf.echoesinink.EchoesInInk;
 import pl.peterwolf.echoesinink.block.ModBlocks;
+import pl.peterwolf.echoesinink.block.HangingPosterBlock;
+import pl.peterwolf.echoesinink.block.LaidPaperBlock;
 import pl.peterwolf.echoesinink.block.entity.InvestigationBlockEntity;
+import pl.peterwolf.echoesinink.block.entity.LaidPaperBlockEntity;
 import pl.peterwolf.echoesinink.config.ModConfig;
+import pl.peterwolf.echoesinink.item.ModItems;
 import pl.peterwolf.echoesinink.progression.InvestigationRole;
 import pl.peterwolf.echoesinink.progression.WorkshopIdentity;
 import pl.peterwolf.echoesinink.progression.WorkshopLayoutPlan;
@@ -169,6 +175,12 @@ public class AbandonedPrintshopPiece extends ScatteredFeaturePiece {
 		placeBlock(level, Blocks.CARPET.white().defaultBlockState(), mx(3, width), 1, 6, box);
 		placeBlock(level, ModBlocks.CARVED_WOODEN_MATRIX.defaultBlockState(), mx(2, width), 2, 3, box);
 		placeBlock(level, Blocks.OAK_STAIRS.defaultBlockState().setValue(StairBlock.HALF, Half.BOTTOM), mx(1, width), 1, 2, box);
+		placeBlock(level, Blocks.CARTOGRAPHY_TABLE.defaultBlockState(), mx(6, width), 1, 8, box);
+		placeLaidPrint(level, box, mx(6, width), 2, 8, Direction.EAST, ModItems.DECORATIVE_WOODCUT);
+		placeLaidPrint(level, box, mx(11, width), 2, 5, Direction.NORTH, ModItems.DAMAGED_ARCHIVE_PAGE);
+		placeHangingPrint(level, box, mx(6, width), 2, 9, Direction.NORTH);
+		placeHangingPrint(level, box, mx(11, width), 2, 1, Direction.SOUTH);
+		placeCobwebs(level, box, new int[][] {{1, 4, 2}, {5, 2, 5}, {7, 3, 8}, {10, 4, 9}});
 		createChest(level, box, random, mx(10, width), 1, 7, RURAL_LOOT);
 	}
 
@@ -208,7 +220,14 @@ public class AbandonedPrintshopPiece extends ScatteredFeaturePiece {
 		placeBlock(level, Blocks.BLAST_FURNACE.defaultBlockState(), mx(14, width), 1, 5, box);
 		placeBlock(level, Blocks.ANVIL.defaultBlockState(), mx(12, width), 1, 6, box);
 		placeBlock(level, Blocks.IRON_BARS.defaultBlockState(), mx(15, width), 2, 5, box);
-		placeBlock(level, ModBlocks.HANGING_POSTER.defaultBlockState(), entranceX + 2, 2, 1, box);
+		placeBlock(level, Blocks.CARTOGRAPHY_TABLE.defaultBlockState(), mx(3, width), 1, 9, box);
+		placeBlock(level, Blocks.SMITHING_TABLE.defaultBlockState(), mx(10, width), 1, 9, box);
+		placeLaidPrint(level, box, mx(3, width), 2, 9, Direction.WEST, ModItems.DECORATIVE_WOODCUT);
+		placeLaidPrint(level, box, mx(10, width), 2, 9, Direction.SOUTH, ModItems.BLANK_ARCHIVE_PAGE);
+		placeHangingPrint(level, box, entranceX + 2, 2, 1, Direction.SOUTH);
+		placeHangingPrint(level, box, mx(2, width), 3, 1, Direction.SOUTH);
+		placeHangingPrint(level, box, mx(15, width), 2, 6, layoutIndex == 0 ? Direction.WEST : Direction.EAST);
+		placeCobwebs(level, box, new int[][] {{1, 4, 1}, {7, 4, 10}, {15, 4, 11}, {12, 3, 2}});
 		createChest(level, box, random, mx(14, width), 1, 10, TOWN_LOOT);
 	}
 
@@ -250,7 +269,13 @@ public class AbandonedPrintshopPiece extends ScatteredFeaturePiece {
 		}
 		placeBlock(level, Blocks.LECTERN.defaultBlockState(), 6, 1, mz(5, depth), box);
 		placeBlock(level, Blocks.CARTOGRAPHY_TABLE.defaultBlockState(), 10, 1, mz(8, depth), box);
+		placeBlock(level, Blocks.BARREL.defaultBlockState(), 12, 1, mz(10, depth), box);
 		placeBlock(level, Blocks.CARPET.white().defaultBlockState(), 8, 1, mz(7, depth), box);
+		placeLaidPrint(level, box, 10, 2, mz(8, depth), Direction.EAST, ModItems.DECORATIVE_WOODCUT);
+		placeLaidPrint(level, box, 12, 2, mz(10, depth), Direction.NORTH, ModItems.DAMAGED_ARCHIVE_PAGE);
+		placeHangingPrint(level, box, 13, 2, mz(2, depth), Direction.WEST);
+		placeHangingPrint(level, box, 1, 3, mz(13, depth), Direction.EAST);
+		placeCobwebs(level, box, new int[][] {{2, 4, mz(2, depth)}, {6, 3, mz(11, depth)}, {12, 4, mz(12, depth)}, {9, 4, mz(4, depth)}});
 		createChest(level, box, random, 4, 1, mz(12, depth), SCHOLARLY_LOOT);
 	}
 
@@ -283,8 +308,13 @@ public class AbandonedPrintshopPiece extends ScatteredFeaturePiece {
 		placeBlock(level, Blocks.IRON_CHAIN.defaultBlockState(), mx(2, width), 3, 3, box);
 		placeBlock(level, Blocks.IRON_BARS.defaultBlockState(), mx(11, width), 1, 7, box);
 		placeBlock(level, Blocks.CARPET.black().defaultBlockState(), mx(6, width), 1, 6, box);
-		placeBlock(level, Blocks.COBWEB.defaultBlockState(), mx(8, width), 3, 9, box);
-		placeBlock(level, ModBlocks.HANGING_POSTER.defaultBlockState(), mx(4, width), 2, 1, box);
+		placeBlock(level, Blocks.CARTOGRAPHY_TABLE.defaultBlockState(), mx(2, width), 1, 8, box);
+		placeBlock(level, Blocks.BARREL.defaultBlockState(), mx(11, width), 1, 4, box);
+		placeLaidPrint(level, box, mx(2, width), 2, 8, Direction.WEST, ModItems.DAMAGED_ARCHIVE_PAGE);
+		placeLaidPrint(level, box, mx(11, width), 2, 4, Direction.NORTH, ModItems.DECORATIVE_WOODCUT);
+		placeHangingPrint(level, box, mx(4, width), 2, 1, Direction.SOUTH);
+		placeHangingPrint(level, box, mx(7, width), 3, depth - 2, Direction.NORTH);
+		placeCobwebs(level, box, new int[][] {{1, 4, 2}, {8, 3, 9}, {11, 4, 10}, {5, 3, 4}});
 		createChest(level, box, random, mx(11, width), -2, 8, BURNED_LOOT);
 	}
 
@@ -448,6 +478,47 @@ public class AbandonedPrintshopPiece extends ScatteredFeaturePiece {
 		BlockPos worldPos = getWorldPos(x, y, z);
 		if (box.isInside(worldPos) && level.getBlockEntity(worldPos) instanceof InvestigationBlockEntity investigation) {
 			investigation.configureWorkshop(workshopId, variant, role);
+		}
+	}
+
+	private void placeLaidPrint(
+		WorldGenLevel level,
+		BoundingBox box,
+		int x,
+		int y,
+		int z,
+		Direction facing,
+		Item item
+	) {
+		placeBlock(
+			level,
+			ModBlocks.LAID_PAPER.defaultBlockState().setValue(LaidPaperBlock.FACING, facing),
+			x, y, z, box
+		);
+		BlockPos worldPos = getWorldPos(x, y, z);
+		if (box.isInside(worldPos) && level.getBlockEntity(worldPos) instanceof LaidPaperBlockEntity paper) {
+			paper.setPage(new ItemStack(item));
+		}
+	}
+
+	private void placeHangingPrint(
+		WorldGenLevel level,
+		BoundingBox box,
+		int x,
+		int y,
+		int z,
+		Direction facing
+	) {
+		placeBlock(
+			level,
+			ModBlocks.HANGING_POSTER.defaultBlockState().setValue(HangingPosterBlock.FACING, facing),
+			x, y, z, box
+		);
+	}
+
+	private void placeCobwebs(WorldGenLevel level, BoundingBox box, int[][] positions) {
+		for (int[] pos : positions) {
+			placeBlock(level, Blocks.COBWEB.defaultBlockState(), pos[0], pos[1], pos[2], box);
 		}
 	}